Remove babel-polyfill and babel-register; not suitable for production
