+++ index.js
@@ -1,163 +1,39 @@
 'use strict'
 
-// Babel ES6 support
-require('babel-register')
-require('babel-polyfill')
-require('./app.js')
+import app from './backend'
+import colors from 'colors'
 
-//
-//                      _oo0oo_
-//                     o8888888o
-//                     88" . "88
-//                     (| -_- |)
-//                     0\  =  /0
-//                   ___/`---'\___
-//                 .' \\|     |# '.
-//                / \\|||  :  |||# \
-//               / _||||| -:- |||||- \
-//              |   | \\\  -  #/ |   |
-//              | \_|  ''\---/''  |_/ |
-//              \  .-\__  '-'  ___/-. /
-//            ___'. .'  /--.--\  `. .'___
-//         ."" '<  `.___\_<|>_/___.' >' "".
-//        | | :  `- \`.;`\ _ /`;.`/ - ` : | |
-//        \  \ `_.   \_ __\ /__ _/   .-` /  /
-//    =====`-.____`.___ \_____/___.-`___.-'=====
-//                      `=---='
-//
-//
-//  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-//
-//      佛祖保佑                永無 BUG
-//
-//
-//
-//
-//                               |~~~~~~~|
-//                               |       |
-//                               |       |
-//                               |       |
-//                               |       |
-//                               |       |
-//    |~.\\\_\~~~~~~~~~~~~~~xx~~~         ~~~~~~~~~~~~~~~~~~~~~/_//;~|
-//    |  \  o \_         ,XXXXX),                         _..-~ o /  |
-//    |    ~~\  ~-.     XXXXX`)))),                 _.--~~   .-~~~   |
-//     ~~~~~~~`\   ~\~~~XXX' _/ ';))     |~~~~~~..-~     _.-~ ~~~~~~~
-//              `\   ~~--`_\~\, ;;;\)__.---.~~~      _.-~
-//                ~-.       `:;;/;; \          _..-~~
-//                   ~-._      `''        /-~-~
-//                       `\              /  /
-//                         |         ,   | |
-//                          |  '        /  |
-//                           \/;          |
-//                            ;;          |
-//                            `;   .       |
-//                            |~~~-----.....|
-//                           | \             \
-//                          | /\~~--...__    |
-//                          (|  `\       __-\|
-//                          ||    \_   /~    |
-//                          |)     \~-'      |
-//                           |      | \      '
-//                           |      |  \    :
-//                            \     |  |    |
-//                             |    )  (    )
-//                              \  /;  /\  |
-//                              |    |/   |
-//                              |    |   |
-//                               \  .'  ||
-//                               |  |  | |
-//                               (  | |  |
-//                               |   \ \ |
-//                               || o `.)|
-//                               |`\\\\) |
-//                               |       |
-//                               |       |
-//  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-//
-//                    耶穌保佑                永無 BUG
-//
-//
-//
-//
-//
-//          ┌─┐       ┌─┐
-//       ┌──┘ ┴───────┘ ┴──┐
-//       │                 │
-//       │       ───       │
-//       │  ─┬┘       └┬─  │
-//       │                 │
-//       │       ─┴─       │
-//       │                 │
-//       └───┐         ┌───┘
-//           │         │
-//           │         │
-//           │         │
-//           │         └──────────────┐
-//           │                        │
-//           │                        ├─┐
-//           │                        ┌─┘
-//           │                        │
-//           └─┐  ┐  ┌───────┬──┐  ┌──┘
-//             │ ─┤ ─┤       │ ─┤ ─┤
-//             └──┴──┘       └──┴──┘
-//  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-//
-//      神獸保佑                永無 BUG
-//
-//
-//
-//
-//
-//                  ___====-_  _-====___
-//            _--^^^#####//      \\#####^^^--_
-//         _-^##########// (    ) \\##########^-_
-//        -############//  |\^^/|  \\############-
-//      _/############//   (@::@)   \\############\_
-//     /#############((     \\//     ))#############\
-//    -###############\\    (oo)    //###############-
-//   -#################\\  / VV \  //#################-
-//  -###################\\/      \//###################-
-// _#/|##########/\######(   /\   )######/\##########|\#_
-// |/ |#/\#/\#/\/  \#/\##\  |  |  /##/\#/  \/\#/\#/\#| \|
-// `  |/  V  V  `   V  \#\| |  | |/#/  V   '  V  V  \|  '
-//    `   `  `      `   / | |  | | \   '      '  '   '
-//                     (  | |  | |  )
-//                    __\ | |  | | /__
-//                   (vvv(VVV)(VVV)vvv)
-//  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-//
-//            神龍保佑                永無 BUG
-//
-//
-//
-//
-//          ▄▄▄██▀▀▀▀███▄
-//        ▄▀▀           ▀█
-//     ▄▄▀               ▀█
-//    █     ▀▄  ▄▀        █
-//    ▐██▄  ▀▄▀▀▄▀  ▄██▀ ▐▌
-//    █▀█ ▀   ▀▀   ▀ █▀  ▐▌
-//    █  ▀▐        ▌▀     █
-//    █                   █
-//     █  ▀▄    ▄▀        █
-//     █           ▄▄    █
-//      █▀██▀▀▀▀██▀      █
-//      █  ▀████▀       █
-//       █            ▄█
-//        ██     █▄▄▀▀ █
-//         ▀▀█▀▀▀▀      █
-//          █            █
-//  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-//
-//            YEE保佑                永無 BUG
-//
-//  ________   ________          ________  ___  ___  ________          ________  ___       ________
-// |\   ___  \|\   __  \        |\   __  \|\  \|\  \|\   ____\        |\   __  \|\  \     |\_____  \
-// \ \  \\ \  \ \  \|\  \       \ \  \|\ /\ \  \\\  \ \  \___|        \ \  \|\  \ \  \     \|___/  /|
-//  \ \  \\ \  \ \  \\\  \       \ \   __  \ \  \\\  \ \  \  ___       \ \   ____\ \  \        /  / /
-//   \ \  \\ \  \ \  \\\  \       \ \  \|\  \ \  \\\  \ \  \|\  \       \ \  \___|\ \  \____  /  /_/__
-//    \ \__\\ \__\ \_______\       \ \_______\ \_______\ \_______\       \ \__\    \ \_______\\________\
-//     \|__| \|__|\|_______|        \|_______|\|_______|\|_______|        \|__|     \|_______|\|_______|
-//
-// source: https://github.com/Lorex/Bugfree.js/blob/master/README.md
+import { dbDev, dbTest, dbProd } from './config'
+import { DatabaseConnector } from './backend/db'
+import { updateTopPosts } from './backend/topPosts'
+import { updateHotPosts } from './backend/hotPosts'
+const port = process.env.PORT || 3000
+const DBConfig = (process.env.NODE_DEV == 'production') ? dbProd : dbDev
+
+;(async () => {
+  try {
+    const info = await DatabaseConnector(DBConfig)
+    console.log(`[${'*'.green}]  Connected to ${info.host}:${info.port}/${info.name}`)
+  } catch (err) {
+    console.log(`[${'!'.red}]  Unable to connect to database`)
+  }
+  await app.listen(port)
+  console.log(`[${'@'.green}]  Server listening on port ${port}`)
+
+})()
+
+// updateTopPosts at first run.
+updateTopPosts()
+
+// updateHotPosts at first run.
+updateHotPosts()
+
+// setInterval to update top posts every 12 hours.
+setInterval(() => {
+  updateTopPosts()
+}, 43200000)
+
+// setInterval to update hot posts every 30 minutes.
+setInterval(() => {
+  updateHotPosts()
+}, 1800000)
